misc/data_generation/requirements.txt
faker>=20.1.0
psycopg2-binary>=2.9.9
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1 